🔨 refactor(app): add cloudflare service
